requirements.txt
# Core Framework
crewai>=0.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.27.0

# LLM Integration
ollama>=0.1.0
openai>=1.3.0
langchain>=0.0.350
langchain-community>=0.0.350
langchain-core>=0.1.0
langsmith>=0.0.1

# Media Processing (Open Source Priority)
openai-whisper
ffmpeg-python>=0.2.0
pydub>=0.25.1
moviepy>=1.0.3

# Vector Embeddings
sentence-transformers>=2.2.2

# Document Generation
reportlab>=4.0.0
python-docx>=0.8.11
fpdf2>=2.7.6
jinja2>=3.1.2

# File Handling
python-multipart>=0.0.6
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
typing-extensions>=4.5.0
